Moves RedisStoreError into separate module and iterates over error_cause in Debug implementation
